requirements.txt
streamlit
openai-whisper
yt-dlp
requests
mediapipe
opencv-python
pandas
tqdm
moviepy==1.0.3
stqdm
Pillow